+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 IMPROVEMENTS:
 * resource/connector_gcp: replaced GCP Deployment Manager API with individual GCP Compute Engine APIs for VM and disk management.
 
+NEW FEATURES:
+* resource/snapmirror: add `delete_destination_volume` option to automatically delete destination volume when destroying snapmirror relationships.
+
 ENHANCEMENTS:
 * Update all the resources documenation by adding `Forces new resource` if the modification is not supported.
-
 
 ## 25.3.0
 NEW FEATURES: